Thread instantiation but no handling yet.
Threads can now be created and can be run through the thread pool
struct (basically a thread manager object). However, these cannot
do not yet execute in a loop (the timeout functionality of each
thread is yet to be coded, plus the thread pool struct is pretty
barebones at the moment).