Reworked thread lib before nstp.
I doubt this is currently a working version of the thread class
but I think the idea behind the implementation is right. I gave each
instance its own mutex to tell it when to run / modify its shared
resource. This way, we have more control over individual threads and
can terminate them individually.