Created the buffer class.
Went through two draft implementations of the buffer class.
The first one was based on a doubly-linked list idea with entries
inside the buffer; however, things got a little hairy and it actually
ended upovercomplicating things. The final idea was to use a simple
array of format strings, with ANSI escape sequences being passed as
parameters to those strings.