Created the grid class.
The grid class uses bits to store information on a 2d interface.
By doing this, we save space and speed up computations on grids.
Also uint32_t was used as this offers the most number of bits we
can use for a single bit string (32). uint64_t was an initial
candidate, but apparently, information is lost when bit shifting
uint64_t beyond 32 places (despite the size of its type; this is
probably due to the architecture of the system).