Implemented random mine generation.
Locations are randomly selected on the grid using the rand() function;
if a selected location is already planted with a mine, the function
moves to the next available cell with a higher index (in other words,
it increments the index of the selected location until a free cell
is reached. Note that it loops to the start whenever its index
overflows).