Big thread utils library update!
A big improvement to the overall design of the thread library was
implemented. Aside from decoupling some of the classes, a new Mutex
class was implemented, and helped abstract the differences in
environment implementations further (since Unix and Windows use
different APIs for handling threads and mutexes). Additionally,
the ThreadPool class was refactored and made much more decoupled
from the Thread class. Many bugs were encountered and fixed (
particularly the sizeof(this->hMutex) => sizeof(*(this->hMutex))).