Added to the event infrastructure; created String utility functions.
The event functions are currently not done, but right now it seems as
if the implementation follows a Monad design pattern (correct me if
I'm wrong). Also, whenever an event fires, it logs its state in an
array so we have a short history of the events that occured.

A String_new() was also implemented and allocated memory for a new string.
free() must be called on any variables that are allocated using this
function.