Refactored the utils.event.h file.
Created an event manager, which works similarly to thread manager
in that it abstracts the classes it manages (especially the methods
tied to those classes). Additionally, events are now implemented through
a priority queue (a linked list in this case) and are resolved one
after the other in a FIFO sequence.