Created draft event handlers.
I also typedefed a pointer to a void function that accepts a char input.
This function represents our template for event handlers. Im also
considering making the event handler a member of the event struct,
so the event can resolve itself without having to get inputs from
elsewhere or what not.