Refactored the way events are implemented.
So I just woke up from a dream where I thought of how to go about this,
and holy sht I think this is pretty good:

Basically, we have a chain of events stored in the event manager. Every
time a new event is created, it is appended to this chain, and it is
up to the manager to resolve these events 1 by 1 as time progresses.

But here's the good part: each event TYPE also has its own eventHandler
chain, so that when resolving an event it simply goes through the chain
associated with its eventType. These chains can be modified by the user
during runtime, much like in JS.