Made the buffer library better.
Basically, it every member of the 2d array stores a reference
to its context. That way, when the value across a row of the context
array becomes different, we know that a context change happens.

ALSO just realized that calling Graphics_getCodeBG means we have to
call free() on the created strings after. MEMLEAKS BABY!