Finished implementing hashmaps.
That was a big fcking headache... a bunch of bugs were encountered
while implementing the HashMap_resize() function, which was made
much more difficult because of the fact that linked lists were used
to handle the collisions, and things got a little tangled, but
everything's okay now.

The next step would be to refactor the existing Manager codes to use
HashMaps to store their children.