Refactored the thread manager to use hashmaps; HashMap_getKeys().
The code for handling threads is now more efficient because it
uses hashmaps instead of an array. Note only that, the code is
also more concise because its easier to look things up by name.

Lastly, I also implemented the HashMap_getKeys() function, which
outputs all the keys in the hashmap to a provided array.