Implemented the AnimationManager.
This is not complete although at the moment, the basic functions of
creating animations and updating them are functional.

I also refactored some of the code to make sure items were not
in the hashmap before instantiating a new object to put into the hash
map (otherwise, we'd have memory leaks from the created objects that
we never ended up putting in the hash map because of duplicate keys).