Fixed SegFault when quitting game; cleaned up main file.
The segmentation fault when quitting the game was apparently
rooted in the ThreadManager_exit() function. Things should work
fine now and the program doesn't hang for a few seconds upon exiting.

Additionally, I editied the main file to be much more descriptive
and helpful to the users. Additional prompts are included, and a
"The game is loading..." has been added so users don't suspect
the program is hanging.