UNICODE CHARACTERS WORK??
So now Unicode characters work. Yes, you heard me, theyre not compatible
with the code base. I had to turn our 2d array of characters in the buffer
into a 3d array: a 2d array of 4-char arrays, where each 4-char represents
a single character. We had to do this because characters outside of ASCII
require more than a single byte for representation.

ALSO it took so long to debug this... my initial idea was to use wchar_t,
but then I realized than only had 2 goddamn bytes... some of the characters
we're using need 3 or 4 (at most 4).

Then i tried using a standard char array, but I stored offsets in a separate
int array so I could compute how much to offset new inserts into the array,
but that failed horribly because every time something was inserted to the
start of the char array and shifts were required, it would have to recompute
the offset array.

Finally, after almost giving up, I resorted to a solution which I thought
was stupid at first but ultimately it actually sufficed.