Fixed Linux asset file parsing bug.
So apparently this had something to do with how differently Windows
and Linux file systems encode new lines. On windows, a /r follows
then /n every time there's a new line, and usually this is just ignored
when reading files I guess. But on Linux, Windows-saved files explicitly
have both and that changes how we have to parse the files on Linux
(in case they were previously saved on Windows).