Created proper page handling.
Now we can switch between pages on cue. The next thing to do would
be to create a eventState class which can store event states for us.
This class will be a member of the each page class, and will be shared
across all page instances. That way, we dont have to pass the instance
down a long chain of callbacks.