Refactored page.utils.h
Holy sht this took so long but it was worth it. Coding the UI
is now more developer friendly.

Basically, the page renders its components through a tree and stores
members of the component class in that tree. Also, page handlers
deal with updating the page now instead of having a separate
animation class.

It may seem monolithic, but honestly structuring the page class this
way keeps things way more simple.

Currently, the original animation has been removed, but this will be put
back in due time.