Sped up Windows framerate; created component dist function.
The component distance function can help us determine whether or
not an animation is close to ending. Also, for some reason, Linux
runs the program faster than Windows. I have yet to figure out
why but will do something about it in the future.

Finally, the animation was slightly improved.