Made sure hashmap duplicates were accounted for; implemented createTextAsset().
HashMap duplicates are not appended by default. However, functions that
implement the HashMap_add() function do not necessarily check for the
success of this operation first. I made sure that all functions that do
this check whether or not a duplicate exists first before resuming
the rest of their instructions; otherwise, we might end up with memory
leaks because a function doesn't end up appending a duplicate to a
hashmap and keeps the object lying around in memory.

Additionally, the createTextAsset() function under the AssetManager
was implemented. This allows us to generate assets with a certain
font family (the font family being a custom font created with Unicode
symbols or something else).