Added icons to menu selection; fixed buffer big.
So we now have icons for each of the pages designed entirely with
Unicode box-drawing characters. These also slide in and out of the
page when selecting where to go next.

Also, I fixed a bug I did not notice; it was probably born when
I decided to integrate Unicode character handling into the code base,
but basically, when a buffer encounters a -x value, it forgets to
increment j2 to shift the starting position on the passed content array.