Fixed a massive bug.
So apparently I forgot to recent the component count associated with
a page whenever I reset that page's component tree. Because of that,
at some point, components stopped getting added to the page during
initting, and becasue of this, the component tree ends up shtting
itself when it tries to construst it.